Discard saved filters if format changes
At least for now, I don't want to deal with backwards
compatibility for locally stored filters. If the format
of the template changes, any stored configs from the
old one should be discarded.

Online filters will work differently. I will ensure to
update any filters in the database whenever necessary.
For locally stored filters, this is a lot harder because
they may be multiple versions behind.

Still, I might add that feature in the future. It's just
not a priority right now, and this fix ensures that the
website remains usable after a change.