Remove currency Hide override defaults
They didn't work because overrides need to state the full name
of the item, otherwise it can't be removed from other categories
and will still show up.
I decided to remove them entirely because the default filter
should also work for leveling a new character in a fresh league,
and you may not have enough transmutes to sell for scrolls in
that situation. Endgame filters can choose a Hidden threshold
above 0 or add the desired overrides manually.