Fix bug breaking filter when editing some numbers
Some of the number text fields were configured to store their
values as string in the config object. This causes the template
renderer to crash.

The fix ensures that the text fields correctly store values as
numbers and additionally adds a filter version upgrade step that
fixes any broken versions created before the fix.