Add League Unique selection
This commit changes the way we process unique prices. Before,
all filtering would happen when scraping from poe.ninja.
Also we would collapse all uniques with the same base type
into a single entry which is the highest cost of that base type.

Now, we allow the user to define custom filter conditions.
This means we need to delay the filtering and we can't collapse
to base types anymore. Moreover, the client needs to know the
individual unique names and prices as well to allow preview of
the settings.